Items will now home in on the player when they are within a certain distance. This functionality is part of the generic item script, so it should work with all items. Updated functionality of text displaying when user hovers over item in game space, it now uses a text mesh instead of a sprite. When a player picks up an item, the item name appears and floats upward, eventually disappearing. This is feedback to the player that they collected an item. Items will home in on the nearest player object.
I made a few changes unrelated to the work item, mainly related to mining, such as the sound effects and the fact you can mine normal asteroids now. Also added background music. Finally, I updated the menu cursor to now be properly positioned in the game window, as it was slightly off before.